muon: per-arch NS dtype default via overridable class attr (sm75 fp16 / sm120 bf16)
Single source of truth: FusedMuon.DEFAULT_NS_DTYPE class attr (sm75=_NS_DTYPE=fp16
for Turing/T4 which has no bf16 tensor cores; sm120 overrides to bf16 for
Blackwell). __init__ takes ns_dtype=None and resolves to self.DEFAULT_NS_DTYPE, so
subclasses/callers reference or override one value instead of scattered literals.
The Frobenius norm is ALWAYS fp32 (fp32 accumulate) regardless of ns_dtype - the
one place low precision hurts NS stability; aurora/normuon norms already fp32.
sm120 gram/symmul fn defaults -> bf16. olm DEF ns_dtype=bf16 (Ada, per v16 wash).

+++ ablate_muon/olm.py
@@ -81,7 +81,7 @@ _DEF = dict(arm="default", seed=0, steps=6000, batch=768, d=128, layers=4, heads
             warmup=500, decay_frac=0.2, min_lr_frac=0.1,               # WSD, same for both opts
             scale_mode="aurora", aurora_k=1, ns_kj=6, coeffs="kj",      # DEFAULT = ns8 (6 KJ) aurora_k1
             #  coeffs: "kj" -> KJ*ns_kj + 2 pin;  "pe" -> Polar-Express PE-8 (8 iters)
-            ns_dtype="fp16", nesterov=True,                             # Muon NS precision + momentum type
+            ns_dtype="bf16", nesterov=True,                             # NS precision: bf16 default (v16: ==fp16, more portable); norms stay fp32
             repulse=0.0, decor=0.0, grad_rep=0.0, xorth=0, niche=0.0,   # mechanism knobs (mech.py)
             scap=0.0, cautious=0.0, grokfast=0.0, gf_alpha=0.98, lookahead=0, la_beta=0.5,
             depth_mix=(0.45, 0.25, 0.15, 0.08, 0.045, 0.025))
